+++ mobile.py
@@ -27,16 +27,16 @@ class EventHandler(base.BaseHandler, tornado.auth.FacebookGraphMixin):
     @tornado.web.asynchronous
     def get(self):
         self.access = 0
-        _eventid = self.get_argument('id')
+        _eventid = self.get_int_argument('id')
         _id = self.get_current_user()['id']
         self.event = self.db.get(
-            "SELECT * FROM willbeout_events WHERE id = %s", int(_eventid))
+            "SELECT * FROM willbeout_events WHERE id = %s", _eventid)
         if not self.event:
             raise tornado.web.HTTPError(404)
         self.places = self.db.query("""select a.id, a.event_id, a.address, a.city, a.name, a.url, a.user_id, a.user_name, count(b.suggestion_id) as friends from willbeout_suggest as a
         LEFT JOIN willbeout_votes as b ON a.id = b.suggestion_id
         WHERE a.event_id = %s
-        GROUP BY a.id ORDER BY friends DESC;""", int(_eventid))
+        GROUP BY a.id ORDER BY friends DESC;""", _eventid)
         if str(self.event['userid']).strip('L') == _id:
             self.access = 1
         self.facebook_request("/me/friends/" + str(self.event['userid']), self